* 增加个无用的PlatformEventHandler::setIsDraggableNodeMousedown。本来想用来判断是否拖拽节点的。但发现经常被其他节点挡住，用处不大 * 修复style="-webkit-app-region: drag"不管用的问题。原因有几个。一个是上个提交的SetCapture，二个是m_draggableRegion没初始化 * 修复最小化后又最大化，layer window没显示完整的问题。原因不清楚，现在规避一下，还原大小后发一次重绘消息

+++ content/browser/NavigationController.cpp
@@ -132,7 +132,7 @@ void NavigationController::insertOrReplaceEntry(const blink::WebHistoryItem& ite
 //             m_items.removeLast();
 //         }
 //         m_items.append(historyItem);
-        if (m_items.size() != 0 && m_currentOffset < m_items.size())
+        if (m_items.size() != 0 && m_currentOffset < (int)m_items.size())
             m_items[m_currentOffset] = historyItem;
         break;
     default:
